refactor(api): apply Rector EARLY_RETURN + DEAD_CODE rules on src/
Auto-refactor Pass 1 of 2 via api/rector-pass1.php.

Applied rule families:
- EARLY_RETURN: control-flow inversions (ternary-to-if where safe) + null named-arg removal
- DEAD_CODE: unused private methods, unused method parameters,
  unused variable assignments, useless @param/@return tags,
  recasting removal

Behavior preserved; src/ scope only. Phase 0 uncommitted files
(rector.php, phpstan.neon, Organization.php namespace fix) and
the 2 temp scaffolding configs (rector-pass1.php, rector-pass2.php)
were intentionally excluded from this commit.

+++ api/src/Controllers/AgentJobApplicationController.php
@@ -171,8 +171,8 @@ class AgentJobApplicationController
             'data' => array_map(fn ($item) => $item->toArray(), $result['items']),
             'meta' => [
                 'total' => $result['total'],
-                'limit' => (int) ($filters['limit'] ?? 50),
-                'offset' => (int) ($filters['offset'] ?? 0),
+                'limit' => $filters['limit'] ?? 50,
+                'offset' => $filters['offset'] ?? 0,
             ],
         ]);
     }
